==> scripts/ppp_Nilu/scripts/01_prepare_table_lefse.Rmd <==
---
title: "Prepare the table for lefse"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
dbListTables(con) %>% data.frame() %>% filter(grepl("otu",.))
```

Select the medium OTU abundance for the 109 patients

```{r}
# the patient table has been updated that there is now a filtering on low counts samples
newptb <- read_csv('../data/clean myeloma OTU by species 02 23 20 filtered.csv')
newpts <- newptb$mrn

# the patient table 
ptb <- read_csv('../data/clean myeloma pttable 01 27 20.csv') %>% 
  select(mrn, Relapse.POD.Status) %>% 
  filter(mrn %in% newpts )

```


```{r}
cts <- get_data_from_query_OTU(0,'counts_ag')


res <- cts %>% 
  filter(sampleid %in% stb$Sample.ID) %>% 
  count(sampleid)
```


```{r}
stb <- read_csv('../data/clean myeloma sdtable 01 27 20.csv')

stb %>% 
  count(Sample.ID)

missing <- paste(setdiff(stb$Sample.ID, res$sampleid),  collapse = '|')

# find the full name of the 3 missing sample

full3 <- cts %>% 
  filter(str_detect(sampleid, missing)) %>% 
  distinct(sampleid)


intersect(stb$Sample.ID, res$sampleid)

# create a column in the stb that is the real full sampleid
stb <- stb %>% 
  mutate(sampleid = if_else(Sample.ID %in% setdiff(stb$Sample.ID, res$sampleid), 
                            str_glue('{Sample.ID}..pool825'),
                            Sample.ID)) %>% 
  filter(mrn %in% newpts)
```


```{r}


# create this taxa column that is like the metaphlan format but with some level of being NA then use the upper level taxa name
annot_all <- get_data_from_query_OTU(2)

annot <- annot_all %>% 
  mutate(ordr =  if_else(ordr == '', str_glue('of_class_{class}'), ordr),
         family =  if_else(family == '', str_glue('of_order_{ordr}'), family),
         genus =  if_else(genus == '', str_glue('of_family_{family}'), genus),
         species =  if_else(species == '', str_glue('of_genus_{genus}'), species)) %>% 
  mutate(taxa = str_glue('k__{kingdom}|p__{phylum}|c__{class}|o__{ordr}|f__{family}|g__{genus}|s__{species}___OTUkey_{key}'))
```

```{r}
CTS <- cts %>% 
  filter(sampleid %in% stb$sampleid) 

# damn I have to calculate of the relative abundance myself damn
CTS_sample_sum <- CTS %>% 
  group_by(sampleid) %>% 
  summarise(total = sum(count))

CTS <- CTS %>% 
  left_join(CTS_sample_sum, by  = 'sampleid') %>% 
  mutate(relab = round(count/total, 5)) %>% 
  left_join(stb %>% 
              select(mrn, sampleid), by = 'sampleid')

# get the medium OTU abundance of the otus
CTS_final <- CTS %>% 
  group_by(otu_key, mrn) %>% 
  summarise(Relab = median(relab))


final <- CTS_final %>% 
  rename(key = otu_key) %>% 
  left_join(annot %>% 
              select(key, taxa), by = 'key') %>% 
  spread(key = 'mrn', value = 'Relab', fill = 0) %>% 
  ungroup() %>% 
  select(-key)
```


```{r}
# reorder the columns 
# make the mrn to be characters
ptb <- ptb %>% 
  mutate(mrn = as.character(mrn))



final <- bind_cols(final[,1],final[,ptb$mrn])

all.equal(colnames(final)[2:ncol(final)], ptb$mrn)
```


```{r}
final %>% 
  write_tsv('../data/otu_mrns_relab_reordered.tsv')

ptb %>% 
  select(Relapse.POD.Status) %>% 
  rename(Relapse_status = Relapse.POD.Status) %>% 
  mutate(Relapse_status = if_else(str_detect(Relapse_status, '^No'), 'No_POD', "Yes_POD")) %>% 
  t %>% 
  write.table('../data/mrns_pheno.txt', sep = '\t', quote = F, row.names = T, col.names = F)
```

```{bash}
cat ../data/mrns_pheno.txt ../data/otu_mrns_relab_reordered.tsv > ../data/test.tsv
```

```{r}
#
mrn <- read_csv('../data/mrn.csv', col_names = F) %>% 
  column_to_rownames('X1') %>% 
  t() %>% 
  as.data.frame() %>% 
  mutate(taxa = as.numeric(as.factor(taxa))) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column('taxa') %>% 
  write_tsv('../data/mrn_for.tsv')
```
