==> scripts/ppp_Nilu/scripts/02_filter_cts_do_lefse.Rmd <==
---
title: "Filter on the count matrix and do lefse"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE) 
library(tidyverse) 
```

Filter on the counts matrix so that At least 10% of samples have that taxa relab > 0.01% (0.0001)

```{r}
counts <- read_tsv('../data/otu_mrns_relab_reordered.tsv') %>% 
  column_to_rownames('taxa') %>% 
  as.matrix()

pass_thre_num <- ceiling(ncol(counts) * 0.1)

counts_fil <- counts[rowSums(counts > 0.0001) >= pass_thre_num, ] 



# write out the filtered matrix
counts_fil %>% 
  as.data.frame() %>% 
  rownames_to_column('taxa') %>% 
  write_tsv('../data/otu_mrns_relab_reordered_filtered.tsv')
```

REMEMBER THAT THE PHENO TABLE SHOULD NOT HAVE ANY SPECIAL CHARACTERS THAT WILL CONFUSE THE LEFSE PROGRAM

```{bash}
# running the lefse with cladogram
~/pipeline/scripts/lefse/lefse_w_cladogram.sh mrns_pheno.txt otu_mrns_relab_reordered_filtered.tsv lefse_res
```
