==> scripts/ppp_Nilu/scripts/03_sort_out_the_frk_samples.Rmd <==
---
title: "uploading newly added samples"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
source('~/db.R')
```

```{r}
new <- read_csv('../data/msk duke sdtable 03 02 20.csv')

old_all <- read_csv('../data/msk duke sdtable 01 18 20.csv')


# the ones that got removed cuz low total counts
rm <- setdiff(old_all$Sample.ID, new$sampleid_clean)

rm %>% 
  data_frame() %>% 
  write_csv('../data/removed_76_from_the_01-18_version.csv', col_names = F)
```


```{r}
old <- bind_rows(read_tsv('~/pipeline/scripts/upload_sample_NCBI/data/metadata-6965072-processed-ok.tsv'),
                 read_tsv('~/pipeline/scripts/upload_sample_NCBI/data/metadata-6987832-processed-ok.tsv')) %>% 
  select(library_ID)


# get the SRR for the 76 removed ones
old %>% 
  filter(library_ID %in% rm) %>% 
  write_csv('../data/low_count_remove_75.csv')
```


```{r}
library(ggvenn)
two_list <- list(old = old$library_ID,
                      new = new$sampleid_clean)
   
ggvenn(two_list) 
```
```{r}
# find out the new ones in new
new_samps <- setdiff(new$sampleid_clean, old$library_ID)

location <- get_data_from_query_OTU(0, 'asv_alpha_diversity_ag')

new_loca <- location %>% 
  filter(sampleid %in% new_samps)
```
```{r}
setdiff( old$library_ID, new$sampleid_clean)
```
