==> scripts/ppp_Nilu/scripts/04_RF_survival.Rmd <==
---
title: "Random Forests for Survival"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 4
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
```

```{r}
library(tidyverse)
library(readxl)
library(randomForestSRC)
library(ggRandomForests)
library(kableExtra)
pretty_print <- function(df) {
  df %>% 
    kable() %>% 
    kable_styling(position = 'left', full_width = F)
}
```

The outcome of interest can be overall survival or Progression-Free-Survival, which Nilu has nicely organized for this cohort.

## The patient data including clinical variables

Need to have a more comprehensive tables of the important clinical variables

```{r}
suv_pt <- read_excel('../data/pttable_anon_040920.xlsx', sheet = 1)
dic <- read_csv('../data/mrn_pid_key07-Apr-2020 12.20.csv')

pt_sub <- suv_pt %>% 
  full_join(dic) %>% 
  select(-pid.anon.v1, -X1, -CleanMyeloma)

# also the table that has the median div in peri period
# get the median and do a summarization
samps <- read_csv('../data/sdtable_anon_040720.csv') %>% 
  group_by(pid.anon.v1) %>% 
  summarise(med_SR = median(simpson_reciprocal)) %>% 
  full_join(dic) %>% 
  select(mrn, med_SR)


# this is the more comprehensive table and subset to have only the 240 included in the survival analysis
pt_all <- read_csv('../data/msk duke pttable 03 02 20.csv')
  
  
pt <- pt_all %>% 
  select(mrn, institution, Sex , Age, disease_cat, Pt.Status_cat, Regimen_cat) %>% 
  filter(mrn %in% pt_sub$mrn) %>% 
  full_join(pt_sub) %>% 
  full_join(samps)

# divide the pt to groups low div and high div 
thre_div <- median(pt$med_SR)

pt <- pt %>% 
  mutate(div_grp = if_else(med_SR > thre_div, 'high', 'low'))

```


```{r}
# change the column types
pt_final <- pt %>% 
  mutate(institution = if_else(institution == 'MSK Auto', T, F),
         Sex = if_else(Sex == 'M', T, F),
         div_grp = if_else(div_grp == 'high', T, F)) %>% 
  select(-mrn, -Pt.Status_cat,-med_SR)

pt_final %>% head(10) %>% pretty_print() 

pt_final %>% 
  summary

```


There can't be any factor strings !!! you need to convert to numeric or logic!!!!!

**VIMP close to zero indicates the variable contributes nothing to predictive accuracy, and negative values indicate the predictive accuracy improves when the variable is misspecified. In the later case, we assume noise is more informative than the true variable. As such, we ignore variables with negative and near zero values of VIMP, relying on large positive values to indicate that the predictive power of the forest is dependent on those variables.**

### For the PFS

```{r}
res_os <- rfsrc(Surv(PFS_landmrk_time, POD_24) ~ . , data = pt_final %>% select(-OS_landmark_time, -DeadorAlive_24), importance = TRUE)

plot(gg_vimp(res_os)) +
  theme(legend.position = c(0.8, 0.2)) +
  labs(fill = "VIMP > 0",
       title = 'Progression free survival')
```

### For the OS

```{r}
res <- rfsrc(Surv(OS_landmark_time, DeadorAlive_24) ~ . , data = pt_final %>% select(-PFS_landmrk_time, -POD_24), importance = TRUE)

#plot(gg_error(res))


plot(gg_vimp(res)) +
  theme(legend.position = c(0.8, 0.2)) +
  labs(fill = "VIMP > 0",
       title = 'Overall survival')
```



## The patient data including only microbiota features

Top 20 most abundant genera count matrix and summarize to get the median abundance for each patient

```{r}
cts_genus <- read.csv('../data/cts_genus.csv', check.names = F, row.names = 1)
dim(cts_genus)

N <- 20
topN_genera <- sort(apply(cts_genus, 1, sum), decreasing = T)[1:N]
topN_ma <- cts_genus[rownames(cts_genus) %in% names(topN_genera),]
```

**Note:** Ruminococcus in two different family

```{r}
# the top 20 most abundant genera
data_frame(genus = names(topN_genera)) %>% 
  arrange(genus) %>% 
  mutate(genus = str_extract(genus, 'f__.+$')) %>% 
  pretty_print()
  
```

```{r}
# the subset samples that are in the peri time period
peri <- read_csv('../data/sdtable_anon_040720.csv')

peri_ma <- topN_ma[,colnames(topN_ma) %in% peri$sampleid_clean] %>% 
  rownames_to_column('genus') %>% 
  gather(key = 'sampleid', value = 'relab', names(.)[2]:names(.)[ncol(.)]) %>% 
  left_join(peri %>% select(pid.anon.v1, sampleid = sampleid_clean)) %>% 
  left_join(dic %>% select(mrn, pid.anon.v1)) %>% 
  group_by(mrn, genus) %>% 
  summarise(med_relab = median(relab)) %>% 
  mutate(genus = str_extract(genus, 'f__.+$')) %>% 
  spread(key = 'genus', value = 'med_relab')

```
 
**VIMP close to zero indicates the variable contributes nothing to predictive accuracy, and negative values indicate the predictive accuracy improves when the variable is misspecified. In the later case, we assume noise is more informative than the true variable. As such, we ignore variables with negative and near zero values of VIMP, relying on large positive values to indicate that the predictive power of the forest is dependent on those variables.** 

### Overall survival

```{r}
# get one table for overall survival
df_OS <- peri_ma %>% 
  column_to_rownames('mrn') %>% 
  as.matrix() 

df_OS_log <- log2(df_OS + 0.0000388) %>% 
  as.data.frame() %>% 
  rownames_to_column('mrn') %>% 
  mutate(mrn = as.numeric(mrn)) %>% 
  full_join(pt_sub %>% select(mrn, OS_landmark_time, DeadorAlive_24)) %>% 
  ungroup() %>% 
  select(-mrn) 
```

```{r}
df_OS <- peri_ma %>% 
  full_join(pt_sub %>% select(mrn, OS_landmark_time, DeadorAlive_24)) %>% 
  ungroup() %>% 
  select(-mrn) 


df_OS %>% 
  write_csv('../data/df_OS_top20_spp.csv')

df_OS_log %>% 
  write_csv('../data/df_OS_log_top20_spp.csv')
```


```{r}
set.seed(111)
micro_OS <- rfsrc(Surv(OS_landmark_time, DeadorAlive_24) ~ . , data = df_OS_log, importance = TRUE)

plot(gg_vimp(micro_OS, nvar = 10)) +
  theme(legend.position = c(0.8, 0.2)) +
  labs(fill = "VIMP > 0",
       title = 'Overall survival') + 
  theme_classic() +
  ggsave('../data/OS_micriobiome.positive.vim.pdf', width = 10, height = 6)



gg_v <- gg_variable(micro_OS, time = 2,  time.labels = '2')
colnames(gg_v)

plot(gg_v, xvar="f__Enterococcaceae|g__Enterococcus", alpha = 0.4) +
  theme(legend.position = 'none') +
  theme_classic() +
  labs(x = 'Enterococcus relative abundance') +
  scale_color_manual(values = c('#00468B','#EC0000')) +
  ggsave('../data/survival_dir_Enterococcus.pdf', width = 5, height = 3)

plot(gg_v, xvar="f__Streptococcaceae|g__Streptococcus", alpha = 0.4) +
  theme(legend.position = 'none') +
  theme_classic() +
  labs(x = 'Streptococcus relative abundance') +
  scale_color_manual(values = c('#00468B','#EC0000')) +
  ggsave('../data/survival_dir_Streptococcus.pdf', width = 5, height = 3)

plot(gg_v, xvar="f__Lachnospiraceae|g__Coprococcus", alpha = 0.4) +
  theme(legend.position = 'none') +
  theme_classic() +
  labs(x = 'Coprococcus relative abundance') +
  scale_color_manual(values = c('#00468B','#EC0000')) +
  ggsave('../data/survival_dir_Coprococcus.pdf', width = 5, height = 3)
 
```



```{r}
# a heatmap to visualize the relationship
library(pheatmap)
relab <- df_OS %>% 
  select(-OS_landmark_time, -DeadorAlive_24) %>% 
  t

colnames(relab) <- paste('pt', seq(1:240), sep = '')

annot <- df_OS %>% 
  select(DeadorAlive_24) %>% 
  mutate(rowname = paste('pt', seq(1:nrow(.)), sep = '')) %>% 
  column_to_rownames('rowname') 

pheatmap(
  relab,
  annotation_col = annot,
  show_rownames = T,
  show_colnames = T,
  filename = '../data/OS.heatmap.jpg',
  height = 10,
  width = 25,
  scale = 'none',
  cluster_rows = T,
  cluster_cols = T
)
```


### Progression free survival

```{r}
set.seed(101)
df_PFS <- peri_ma %>% 
  full_join(pt_sub %>% select(mrn, PFS_landmrk_time, POD_24)) %>% 
  ungroup() %>% 
  select(-mrn)


micro_PFS <- rfsrc(Surv(PFS_landmrk_time, POD_24) ~ . , data = df_PFS, importance = TRUE, tree.err = TRUE)

plot(gg_vimp(micro_PFS, nvar = 5)) +
  theme(legend.position = c(0.8, 0.2)) +
  labs(fill = "VIMP > 0",
       title = 'Progression Free Survival')+
  theme_classic() +
  ggsave('../data/PFS_micriobiome.positive.vim.pdf', width = 10, height = 6)


```
